+++ client/src/pages/Member.tsx
@@ -1,7 +1,7 @@
 import MemberInfo from '../components/member/MemberInfo/MemberInfo';
 import Ad from '../components/member/Ad';
 import MemberContents from '../components/member/memberContents/MemberContents';
-import ProfileModal from '../components/member/ProfileModal';
+import ProfileModal from '../components/member/profileModal/ProfileModal';
 import { useRecoilValue } from 'recoil';
 import { profileModalState } from '../recoil/atoms/Atoms';
